requirements.txt
tensorflow>=2.0
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
optuna>=2.0
numpy>=1.19
pandas>=1.1
scikit-learn>=0.24
nltk>=3.5
PyYAML>=5.4
keras-cv>=0.4.0
transformers>=4.0
huggingface-hub>=0.12.0
